backend/requirements.txt
Flask
flask-SQLAlchemy
flask-cors
simplekml
PyGithub